backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eduardo da Silva",email = "eduardo.silva@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "drf-spectacular>=0.26.4",
    "python-dotenv>=1.0.0",
    "netifaces>=0.11.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.2.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"